Kubernetes and Rancher deployment guide and files
The start of a deployment guide for Folio on Kubernetes and Rancher 2.0. Currently using Kubernetes v1.10.5 and Rancher v2.0.6. Very basic now, the intent being for this to become more in-depth as notes and deployments mature.

Use the included Dockerfiles and YAML at your own risk! The Dockerfiles are custom images created at TAMU and the YAML was exported from Rancher 2.0 GUI.

Substituted generic ## and <url/ip info> in the code for config at TAMU.